+++ projects/ion/src/lib/date-picker/date-picker-basic/date-picker.component.scss
@@ -1,4 +1,4 @@
-@import '../../styles/index.scss';
+@import '../../../styles/index.scss';
 
 * {
   font-family: 'Source Sans Pro', sans-serif;
