--- projects/ion/src/lib/date-picker/date-picker.component.html
@@ -1,129 +1,0 @@
-<div class="date-container" id="date-container">
-  <div class="date-content">
-    <div
-      *ngIf="!isDateRange"
-      class="field-date"
-      (click)="setFocusOnClickInput('input-date')"
-    >
-      <input
-        id="input-date"
-        type="text"
-        [value]="dates.date.dateLabel || ''"
-        placeholder="Data"
-        readonly
-      />
-    </div>
-    <div class="date-ranges" *ngIf="isDateRange">
-      <div
-        class="field-date"
-        (click)="setFocusOnClickInput('input-start-date')"
-      >
-        <input
-          id="input-start-date"
-          type="text"
-          [value]="dates.startDate.dateLabel || ''"
-          placeholder="Data de início"
-          readonly
-        />
-      </div>
-      <div class="separator" (click)="setFocusOnClickInput('input-start-date')">
-        -
-      </div>
-      <div class="field-date" (click)="setFocusOnClickInput('input-end-date')">
-        <input
-          id="input-end-date"
-          type="text"
-          [value]="dates.endDate.dateLabel || ''"
-          placeholder="Data de fim"
-          readonly
-        />
-      </div>
-    </div>
-    <div class="icons" (click)="actionClickIcon()">
-      <ion-icon
-        *ngIf="!isVisibleIconClose"
-        id="icon"
-        type="calendar"
-        color="#AEB2BD"
-        size="16"
-      ></ion-icon>
-      <ion-icon
-        *ngIf="isVisibleIconClose"
-        id="icon"
-        type="close"
-        color="#505566"
-        size="16"
-      ></ion-icon>
-    </div>
-  </div>
-</div>
-
-<div id="calendar">
-  <div class="calendar-dropdown" id="calendar-dropdown">
-    <div class="header">
-      <ion-button
-        label="<<"
-        type="ghost"
-        (ionOnClick)="previousYear()"
-      ></ion-button>
-      <ion-button
-        label="<"
-        type="ghost"
-        (ionOnClick)="previousMonth()"
-      ></ion-button>
-      <label id="month-year">
-        {{ monthYear }}
-      </label>
-      <ion-button
-        label=">"
-        type="ghost"
-        (ionOnClick)="nextMonth()"
-      ></ion-button>
-      <ion-button
-        label=">>"
-        type="ghost"
-        (ionOnClick)="nextYear()"
-      ></ion-button>
-    </div>
-    <div class="time-interval" *ngIf="isDateRange">
-      <ion-divider></ion-divider>
-      <div class="content">
-        <label>Período:</label>
-        <ion-chip
-          label="7 dias"
-          [selected]="selectedChips.sevenDays"
-          (events)="chipEvents($event, 7)"
-        ></ion-chip>
-        <ion-chip
-          label="15 dias"
-          [selected]="selectedChips.fifteenDays"
-          (events)="chipEvents($event, 15)"
-        ></ion-chip>
-        <ion-chip
-          label="21 dias"
-          [selected]="selectedChips.twentyOneDays"
-          (events)="chipEvents($event, 21)"
-        ></ion-chip>
-      </div>
-      <ion-divider></ion-divider>
-    </div>
-    <div class="week-days" id="week-days">
-      <span *ngFor="let weeKDay of getWeekDaysElementStrings()">
-        {{ weeKDay }}
-      </span>
-    </div>
-    <div class="month-days" id="month-days"></div>
-    <div class="footer" *ngIf="isDateRange">
-      <ion-button
-        label="Cancelar"
-        type="ghost"
-        (ionOnClick)="clearCalendar(true)"
-      ></ion-button>
-      <ion-button
-        label="Pronto"
-        [disabled]="isDisabledConfirmButton"
-        (ionOnClick)="emmitEvent()"
-      ></ion-button>
-    </div>
-  </div>
-</div>
